New custom linker script for the bootloader based off the avr5.x script included with avr-libc.
This new linker script puts the code in the ATmega328's 4KB bootloader section at the high-end of flash and sets the correct sizes for the RAM and EEPROM spaces.  We'll soon define SFR addresses in the linker script for access in the code.  The main.c module was used to test this possibility.