Added watchdog, shared data, and fixed deviceprog module, among other things.
The processor can be reset using the watchdog and either land in the bootloader or
restart the application.  This is done with a new linker output section called
.shareddata that is not initialized on startup.  The section allows data to be shared
by the application and bootloader that is retained across device resets.  A function
tests whether or not this data is valid.  This was tested and does appear to work.  Also,
flash programming seems to work since I could program a dummy appinfo_t struct to
flash, even when it crosses page boundaries.