The device will no longer start up in the bootloader when first powered on.
If the device was reset by the watchdog, we check a flag that says whether or not we
should restart the app or stay in the bootloader. If it was reset by a different source,
such as the external reset or power-on reset, then we'll load the app since that's the
expected behavior.