Working on a serial command interface.
The interface will be the means of communication between the PC
and the device.  It is set up so that most commands will be text-
based, so that you can use them through a terminal like GTKTerm or
TeraTerm.  The only exception so far is when writing program data
to the bootloader; a binary interface will probably be used for
speed and space concerns.

Right now, there are bootcmd.c.h and cmd.c.h.  I thought that the
latter would end up too big for the bootloader, so the former were
created.  However, it might turn out that I can go back to a
single module.  I'll have to see what the space requirements are
like.