More work on serial command interface.
Moved back to a common codebase for the interface code.  Added
command for printing out help function for all of the functions
as well as commands to restart the app or bootloader.  I may move
those out of common code later since there's conditional macros
for those anyway.

Also will need to make a way to choose if help strings are in RAM
or flash.