Numerous improvments to bootloader.
Removed the appinfo_t structure that was going to store the app
CRC and number of pages within application space.  Instead, I no
longer care about the number of pages and the app's CRC will be
stored in bootloader space.  This simplifies code quite a bit as I
was able to remove quite a bit of stuff from sharedmem.c.h and
deviceprog.c.h, as well as the linker script.

Also implemented all of the needed interface commands for the
bootloader and added a new module (localcmd.c) for them; however,
the bootloader is now too big for the 4KB limit.  I'll have to
slim it down.