Bootloader now actually fits in bootloader space.
Removed use of strtoul() in PageProgram_CMD and used atol() and strchr()
instead.  Also shortened the command help strings a bit to save a tiny
bit of space.

Then added function to "rewind" all of the settings that
the bootloader sets up when starting up.  This is called when the
jump to app command is run.