The bootloader actually works now!
I was able to download a small application with a test script in pjcbootloader.py after making some fixes to both the Python and the bootloader code.

Also, the bootloader is a bit smaller now.  The very last page of flash is dedicated to just the app validation key and CRC, which simplified a little of the bootloader code.  The old code was the cause of an update problem, so that was another reason to remove it.