Multiple changes to bootloader code, including change in how updates are performed.
Renamed system.h to appcfg.h since that seems more appropriate for what the file does (did this for app as well).
Made filling the AVR's flash write latches a separate function from programming a page.
Erasing EEPROM now manually uses the AVR erase-only mode and erases only cells that are not already clear.  This no longer uses the AVR libc function for writing to EEPROM since that always uses erase-and-write mode.
Loading page data is now a separate serial command from programming the page.  This makes programming easier since one no longer has a small window in which to write binary data over the serial port to the device.  The new command takes a page offset, checksum, and a hex string representing page data.
Increased the size of the command buffer to 64 bytes to accomodate the new command.
The cmdinfo_t structure now requires that the help string be non-NULL.  This allowed a simplification of the help text state in the Cmd_ProcessInterface() function.