Multiple software changes to work with the most recent bootloader updates.
Added new base class for the PJCBootloader and future PJCApplication classes which handles common commands and executing commands by sending them, reading the response, and parsing it into a useable value.
Added new exception classes that can be raised while trying to execute a serial command.  These will provide better diagnostic info to the software and remove the requirement that command functions check for a lack of response or bad data or whatever.
Modified the PJCBootloader class to use the PJCInterface base class, which cut down the amount of code considerably.  Also modified the class to work with the most recent bootloader's changes to the update sequence.
The CommThread was updated for the most recent bootloader.
There is now a text log widget on the main window that logs text from the CommThread.  Previously, text was output to stdout.
Added an MD5 Sum field to the update page that will display the MD5 hash of the selected hex file.