Add commands to the app and fix an issue in common code.
Clear watchdog when grabbing new command data in Cmd_ProcessInterface.  Weird errors were popping up while testing the software and this seemed to be the issue.
Add a few basic serial commands to the application and have it run the command interface.  This now allows the app to work with the software, which assumes that the app can jump to the bootloader and return version information.