Added new signal/slot manager class and fixed the worker thread.
The new ConnectionManager class makes it easier to pair signals with slots by conecting signals and slots with the same given name automatically.
It also turns out that you are not supposed to inherit from QThread but rather create your separate class and move it to a new QThread using the moveToThread() method.  See http://labs.qt.nokia.com/2010/06/17/youre-doing-it-wrong/.